Separating style from components and added README.md
