+++ infiniteScroll/README.md
@@ -1,0 +1,11 @@
+## Infinite Scroll
+
+App que consulta a TV Shows API (https://www.episodate.com/api). Traz 20 shows mais populares e ao chegar ao fim da página carrega mais 20 shows.
+Utiliza React Hooks, TypeScript, axios.
+
+![App screenshot](./demo/demo.gif)
+
+Para rodar um projeto deve-se seguir os seguintes passos:
+- Para instalar as dependências necessárias: `yarn install`
+- Para rodar o app: `yarn start`
+- Para instalar o app em um device android: `yarn android`
