+++ infiniteScroll/src/components/header/header.styles.ts
@@ -1,0 +1,16 @@
+import {StyleSheet} from 'react-native';
+import colors from '../../styles/colors';
+
+const styles = StyleSheet.create({
+  header: {
+    backgroundColor: colors.primaryDark,
+  },
+  title: {
+    fontSize: 30,
+    fontWeight: 'bold',
+    color: 'white',
+    margin: 16,
+  },
+});
+
+export default styles;
