+++ infiniteScroll/src/components/home/home.styles.ts
@@ -1,0 +1,10 @@
+import {StyleSheet} from 'react-native';
+import colors from '../../styles/colors';
+
+const styles = StyleSheet.create({
+    container: {
+      backgroundColor: colors.background,
+    },
+  });
+
+export default styles;
