+++ infiniteScroll/src/components/showsList/showsList.styles.ts
@@ -1,0 +1,9 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+  },
+});
+
+export default styles;
